move "execute python script" from grizzly.steps.scenario to grizzly.steps
makes it possible to use steps in background section in feature-file.
